// KInvFront/src/app/ServiceSwagger/model/empleadoListAllDateDto.ts
/**
 * Backend
 * Descripción
 *
 * OpenAPI spec version: 0.0.1
 * Contact: willypanza11@gmail.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface EmpleadoListAllDateDto { 
    apellidos?: string;
    cedula?: string;
    direccion?: string;
    email?: string;
    estado?: string;
    estadoVacunacion?: string;
    fechaNacimiento?: Date;
    idEmpleado?: number;
    idUsuario?: number;
    nombreUsuario?: string;
    nombres?: string;
    password?: string;
    salario?: number;
    tipo?: string;
}